build: 升级依赖版本并迁移 Compose 编译器配置
- 升级 Kotlin 至 2.0.21，AGP 至 8.13.2，Gradle 至 8.13
- 迁移 Compose 编译器：移除已废弃的 composeOptions 配置，
  改用 org.jetbrains.kotlin.plugin.compose 插件
- 同步升级 KSP、Compose BOM、Room、Koin 等核心依赖
- 升级百度地图 SDK 至 8.1.0、高德定位至 11.2.000，
  并适配百度地图 API 参数类型变更（Int → Float）

+++ app/src/main/java/com/suseoaa/locationspoofer/ui/components/AppMapView.kt
@@ -336,7 +336,7 @@ class BaiduMapControllerImpl(
         map.addOverlay(
             com.baidu.mapapi.map.PolylineOptions()
                 .color(colorInt)
-                .width(width.toInt())
+                .width(width)
                 .points(latLngList)
         )
     }
@@ -346,7 +346,7 @@ class BaiduMapControllerImpl(
                 .center(com.baidu.mapapi.model.LatLng(lat, lng))
                 .radius(radius.toInt())
                 .fillColor(fillColorInt)
-                .stroke(com.baidu.mapapi.map.Stroke(strokeWidth.toInt(), strokeColorInt))
+                .stroke(com.baidu.mapapi.map.Stroke(strokeWidth, strokeColorInt))
         )
     }
     override fun addMarker(lat: Double, lng: Double, title: String, type: MarkerType): AppMapMarker {
